+++ src/services/passwordServices.ts
@@ -14,3 +14,21 @@ export async function getHashedPassword(res: Response, plainPassword: string): P
     return '';
   };
 };
+
+export async function compareHashedPassword(res: Response, plainPassword: string, hashedPassword: string): Promise<boolean> {
+  try {
+    const isMatch: boolean = await bcrypt.compare(plainPassword, hashedPassword);
+
+    if (!isMatch) {
+      return false;
+    };
+
+    return true;
+
+  } catch (err: any) {
+    console.log(err);
+    res.status(500).json({ success: false, message: 'Internal server error.' });
+
+    return false;
+  };
+};